Check if this rpmdevtools is installed
rpmdev-setuptree is provided by rpmdevtools.
So check that rpmdevtools is installed before trying to use.

+++ scripts/create-rpms.sh
@@ -7,6 +7,13 @@ if [ $? -eq 1 ]; then
 	exit 1
 fi
 
+#check if rpmdevtools is installed
+yum list installed rpmdevtools
+if [ $? -eq 1 ]; then
+	echo "'rpmdevtools' package not installed.. exiting"
+	exit 1
+fi
+
 rm -rf ~/rpmbuild
 rpmdev-setuptree
